requirements.txt
langchain-google-genai

langchain_core

python-dotenv

streamlit

langchain_community

langserve